Added solution
Implemented the following homework:

Create an HTML page that should have a field for entering a nickname, five control buttons (start game, show the best result of this session and local, clear the best result of the session and local), and one button for scoring.

Implemented a check of the nickname for its presence, for a minimum length of 3 characters and for the fact that the initial phrase is not a nickname.

Additionally added tracking of points scored and the remaining time.